requirements.txt
aiogram==3.13.1
openai==1.51.2
google-generativeai==0.8.3
python-dotenv==1.0.1
requests==2.32.3
aiohttp>=3.9.0,<3.11.0